manim-paper-agent/requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.1
manim>=0.18.1
rich>=13.7.0
jsonschema>=4.21.1